Move `Follow` command
