+++ wcfsetup/install/files/lib/action/UserFollowAction.class.php
@@ -11,7 +11,7 @@ use wcf\data\user\User;
 use wcf\http\Helper;
 use wcf\system\exception\IllegalLinkException;
 use wcf\system\exception\PermissionDeniedException;
-use wcf\system\user\command\Follow;
+use wcf\command\user\Follow;
 use wcf\system\user\command\Unfollow;
 use wcf\system\WCF;
