+++ wcfsetup/install/files/lib/data/user/follow/UserFollowAction.class.php
@@ -10,7 +10,7 @@ use wcf\system\cache\runtime\UserProfileRuntimeCache;
 use wcf\system\exception\PermissionDeniedException;
 use wcf\system\exception\UserInputException;
 use wcf\system\user\activity\event\UserActivityEventHandler;
-use wcf\system\user\command\Follow;
+use wcf\command\user\Follow;
 use wcf\system\user\command\Unfollow;
 use wcf\system\user\GroupedUserList;
 use wcf\system\user\storage\UserStorageHandler;
